+++ back/.env
@@ -1,4 +1,4 @@
 DB_HOST=localhost
-DB_USER=JJChan96
-DB_PASSWORD=1431
+DB_USER=sila
+DB_PASSWORD=11111111
 DB_DATABASE=team4
