--- back/routes/board/board_qna/QnA_deleteController.js
@@ -1,25 +1,0 @@
-const pool = require('../../../db.js').pool
-
-exports.QnA_del = async (req,res)=>{
-    console.log('요청 옴')
-    const {idx} = req.body
-    console.log(idx)
-    const sql = `delete from QnA where idx=?`
-    const param = [idx]
-    try {
-        const [result] = await pool.execute(sql,param)
-        const response = {
-            result,
-            errno:0
-        }
-        res.json(response) 
-    } 
-    catch (e) {
-        console.log(e.message)
-        const response = {
-            errormsg: e.message
-        }
-        
-        res.json(response)  
-    }
-}
