+++ dailyrecord_lsj.md
@@ -192,9 +192,12 @@ done
 
 삭제가 되면 글 리스트로 보내준다.
 
+1.3 글 수정
 
 
-1.3 댓글 기능
+
+
+1.4 댓글 기능
 
 댓글을 써서 서밋 한다 > DB에 추가한다.
