+++ front/views/board/QnA/QnA.html
@@ -63,11 +63,11 @@
                 withCredentials:true
             }
 
-            const response = await axios.post('http://localhost:4000/api/board/QnA', data, option)
+            const response = await axios.post('http://localhost:4000/api/board/QnA/write', data, option)
 
             if(response.data.errno === 0) {
                 alert('QnA 작성이 완료되었습니다.')
-                location.href='/board/qna_list'
+                location.href='/board/qna'
             }
             else {
                 alert('QnA 작성중 문제가 발생하였습니다.')
