+++ front/views/board/QnA/QnA_list.html
@@ -5,8 +5,6 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <script src="https://unpkg.com/axios/dist/axios.min.js"></script>
-
 </head>
 <body>
     <h1><a href='/'>logo</a></h1>
